add jobs and cronjobs for syncing chart repos
As part of moving the chart repository syncing outside of the main
Monocular API process, we will use a CronJob to sync chart repositories
on a schedule. The list of repositories to sync will be defined by the
value in this chart. A Job is used for the initial sync at install time.

Signed-off-by: Adnan Abdulhussein <adnan@bitnami.com>