change charts list component to extract repos from list of charts
The old Monocular API supported a repos endpoint that returned a list of
configured repos, but this endpoint is not supported by the chartsvc.
Instead of reimplementing it, we can extract the list of repos from the
charts list response client-side. This updates the chart list component
to always fetch all charts (regardless of a selected repo) allowing it
to extract the full list of repos to show. The full charts list is then
filtered by the selected repo to preserve the same functionality as the
/api/chartsvc/v1/charts/{repoName} endpoint previously used.

Signed-off-by: Adnan Abdulhussein <adnan@bitnami.com>