enable ability to index new chart data for the same chart version
Previously, chart-repo skipped indexing any chart versions previous
indexed based on the repo, name and version. However, if a user modifies
a chart version without bumping a new version, the changes are not taken
into account. This changes chart-repo to skip based on the above AND the
digest of the chart package so that packages with a new digest are
correctly updated.

See also https://github.com/kubeapps/kubeapps/issues/659

fixes #555

Signed-off-by: Adnan Abdulhussein <adnan@bitnami.com>