fetch repos from server
This adds a new repos.json file that holds the information about
what repositories are being synced, allowing us to remove the current
hack that parses the full charts list response for the list of
repositories. With this, we can once again use the /charts/{repo}
endpoint to fetch charts from a single repo.

Signed-off-by: Adnan Abdulhussein <adnan@bitnami.com>