[chart-repo] Support case insensitive readme file in charts (#639)
Fix #638. Sometimes the chart creator uses the name 'README.md',
'readme.md' or 'Readme.md' in the chart tarball. This patch adds
support for case insensitive filenames in the chart tarball.
We assume any files in the same folder in the chart tarball won't
have the same file name when case ignored, and it should be true.

Signed-off-by: Jesse Hu <seizethetime@163.com>