Make search url accessible without chartsvc in the name
The search URL is going public as part of Helm. The Helm CLI
calls it for search. The name of the internal service should
not be part of the semi-public API in case we make changes
in the future.

This change makes both the old and new paths accessible via
ingress. The Helm v3 betas are calling the existing path so we
do not want to remove it and break their experience.

Signed-off-by: Matt Farina <matt@mattfarina.com>